Close idle connections on server error
In order to ensure we do a new DNS lookup when calls to the server are
failing, close idle connections in this case.  This ensures we don't end
up making requests to a broken server that's been taken out of the DNS
pool because it's unhealthy.

Also, discard unused Body to ensure re-use in the typical case where we
don't want to close the connection.